fixed delete post by blogger. added check if post exists

+++ src/features/api/bloggers/blogs/application/useCases/deletePost.useCase.ts
@@ -1,5 +1,6 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { PostSqlRepository } from '../../../../../entities/postgres/postSql.repository';
+import { NotFoundException } from '@nestjs/common';
 
 export class DeletePostCommand {
   constructor(public id: string) {}
@@ -10,6 +11,8 @@ export class DeletePostUseCase implements ICommandHandler<DeletePostCommand> {
   constructor(private postRepo: PostSqlRepository) {}
 
   async execute(command: DeletePostCommand): Promise<any> {
+    const post = await this.postRepo.getPostById(command.id);
+    if (!post) throw new NotFoundException();
     return this.postRepo.deletePost(command.id);
   }
 }
